index.html
Privacy policy
Striking Spear as developer do NOT collect any kind of information such as personal information or personal identifiable information. We do not own any servers to store any information. 
Camera permission - Camera permission is needed for infrared detector feature. This feature needs to open device camera to detect IR lights so this permission is needed for the app

Ads - This app is ads